# docs/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Social Network for Music API",
    "description": "Social Network for Music REST API",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get information about the user with the given id (AUTH required)",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/search/users/{q}": {
      "get": {
        "tags": [
          "Search"
        ],
        "summary": "Get user by query from the database",
        "description": "",
        "parameters": [
          {
            "name": "q",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/search/playlist/{q}": {
      "get": {
        "tags": [
          "Search"
        ],
        "summary": "Get playlist by query from the database",
        "description": "",
        "parameters": [
          {
            "name": "q",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/playlist/": {
      "get": {
        "tags": [
          "Playlist"
        ],
        "summary": "Gets playlists created or saved by the user with the given id (AUTH required)",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/playlist/{id}": {
      "get": {
        "tags": [
          "Playlist"
        ],
        "summary": "Get playlist by id (AUTH required)",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}